Restructure layout
